+++ packages/clerk-js/src/core/resources/internal.ts
@@ -43,4 +43,4 @@ export * from './UserOrganizationInvitation';
 export * from './Verification';
 export * from './Web3Wallet';
 export * from './Waitlist';
-export * from './ApiKey';
+export * from './APIKey';
